Correct : move Component Directory
